add linux-ubuntu compile

+++ Main.py
@@ -37,6 +37,7 @@ from libpy import TypeTrans
 from libpy import GUIMain
 
 DEBUG = 1
+Bounding = 1
 
 epision = 4;
 minS = 25;
@@ -77,12 +78,18 @@ def ReadData():
 
 def main(Model, FileLoc, FileName):
 	os.system("rm -rf mainpy")
-	os.system("gcc -I ./lib mainpy.c -o mainpy")
+	if Init.SystemJudge() == "Linux":
+		os.system("gcc -I ./lib mainpy.c -o mainpy -lm")
+	elif Init.SystemJudge() == "MacOS":
+		os.system("gcc -I ./lib mainpy.c -o mainpy")
+	else:
+		print("Sorry, this system cannot be used under Windows plantform")
 
 	OldImg = cv2.resize(np.array(Image.open(FileLoc[0]).convert("L")), (Owidth, Oheight))
 	TypeTrans.Img2Dat("df", "", "Input/Inp2.dat", OldImg)
 	if Model == "test":
-		print(FileName[0] + ":  ")
+		print(FileName[0] + ":\t")
+		model = pickle.load(open("Output/model.m", 'rb'))
 	TrainX = []
 	TrainY = []
 	for kase in range(1, len(FileLoc)):
@@ -101,7 +108,7 @@ def main(Model, FileLoc, FileName):
 		if not os.path.exists("Output/HogDEMOResult.dat"):
 			continue
 		ClusImg, HogDescriptor, clusMax = ReadData()
-		
+
 
 
 		if DEBUG:
@@ -118,15 +125,16 @@ def main(Model, FileLoc, FileName):
 						imgclus[i][j] = iro[ClusImg[i][j] % 12]
 			imageio.imwrite("Output/Img3.png", imgclus)
 
+			TypeTrans.Dat2Img("ff", "Output/DiffImg.dat", "Output/Img4.png", [])
+
 
 
 		if Model == "train":
-			os.system("rm -rf Output/Img1.png")
-			os.system("rm -rf Output/Img2.png")
-			imageio.imwrite("Output/Img1.png", cv2.cvtColor(OldImg, cv2.COLOR_GRAY2BGR))
-			imageio.imwrite("Output/Img2.png", cv2.cvtColor(NewImg, cv2.COLOR_GRAY2BGR))
-			
-
+			if not DEBUG:
+				os.system("rm -rf Output/Img1.png")
+				os.system("rm -rf Output/Img2.png")
+				imageio.imwrite("Output/Img1.png", cv2.cvtColor(OldImg, cv2.COLOR_GRAY2BGR))
+				imageio.imwrite("Output/Img2.png", cv2.cvtColor(NewImg, cv2.COLOR_GRAY2BGR))
 			
 			Sign = GUIMain.DBSCANGUI(ClusImg, clusMax, Oheight, Owidth)
 
@@ -141,26 +149,66 @@ def main(Model, FileLoc, FileName):
 					TrainY.append(0)
 
 
+
 		if Model == "test":
-			model = pickle.load(open("Output/model.m", 'rb'))
 			TestX = []
 			for i in range(1, len(HogDescriptor)):
 				TestX.append(HogDescriptor[i])
+			if len(TestX) == 0:
+				print()
+				continue
 			Result = model.predict(TestX)
 			
 			NotPrint = True
-
+			BoundingClus = []
+			print(clusMax, end = "\t")
 			for i in range(0, len(Result)):
-				if Result[i] == 1:
-					print("WARNING!!!!")
-					ImgName = "Output/WARNING/img" + str(kase) + ".png" 
-					imageio.imwrite(ImgName, cv2.cvtColor(OldImg, cv2.COLOR_GRAY2BGR))
-					NotPrint = False
-					break
-			
+				if not Bounding:
+					if Result[i] == 1:
+						print("WARNING!!!!")
+						ImgName = "Output/WARNING/" + FileName[kase] + ".png" 
+						imageio.imwrite(ImgName, cv2.cvtColor(OldImg, cv2.COLOR_GRAY2BGR))
+						NotPrint = False
+						break
+				else:
+					if Result[i] == 1:
+						print("W", end = "\t")
+						BoundingClus.append(i + 1)
+
+			if Bounding:
+				RGBImg = cv2.cvtColor(NewImg, cv2.COLOR_GRAY2BGR)
+				BoxData = [[99999999, 99999999, -1, -1] for n in range(len(BoundingClus))]
+				for i in range(0, len(ClusImg)):
+					for j in range(0, len(ClusImg[i])):
+						if ClusImg[i][j] == 0:
+							continue
+
+						for k in range(0, len(BoundingClus)):
+							if ClusImg[i][j] == BoundingClus[k]:
+								BoxData[k][0] = min(i, BoxData[k][0])
+								BoxData[k][1] = min(j, BoxData[k][1])
+								BoxData[k][2] = max(i, BoxData[k][2])
+								BoxData[k][3] = max(j, BoxData[k][3])
+
+				for k in range(0, len(BoxData)):
+					for i in range(BoxData[k][0], BoxData[k][2] + 1):
+						RGBImg[i][BoxData[k][1]][1] = 255
+						RGBImg[i][BoxData[k][3]][1] = 255
+
+					for j in range(BoxData[k][1], BoxData[k][3] + 1):
+						RGBImg[BoxData[k][0]][j][1] = 255
+						RGBImg[BoxData[k][2]][j][1] = 255
+
+				print()
+				imageio.imwrite("tmpout/" + FileName[kase] + ".png" , RGBImg)
+
+
+
 			if NotPrint:
 				print()
-			input()
+			
+			
+
 
 	if Model == "train":
 		FileName = "Output/SaveTrain.dat"
@@ -173,8 +221,23 @@ def main(Model, FileLoc, FileName):
 		File.write(Str)
 		File.close()
 
+		FileName = "Output/CTrain.dat"
+		File = open(FileName, "w")
+		Str = ""
+		for i in range(0, len(TrainX)):
+			if TrainY == 1:
+				Str += str(TrainY[i])
+			else:
+				Str += str("-1")
+			Str += " "
+			for j in range(0, len(TrainX[i])):
+				Str += str(j + 1) + ":" + str(TrainX[i][j]) + " "
+			Str += "\n"
+		File.write(Str)
+		File.close()
 
 	return
+
 
 
 def trainff(FileLoc):
@@ -198,13 +261,43 @@ def trainff(FileLoc):
 
 	pickle.dump(clf, open("Output/model.m", 'wb'))
 
+
+
+	if not os.path.exists("Output/CTrain"):
+		FileName = "Output/CTrain.dat"
+		File = open(FileName, "w")
+		Str = ""
+		for i in range(0, len(TrainX)):
+			if TrainY[i] == 1:
+				Str += str(TrainY[i])
+			else:
+				Str += str("-1")
+			Str += " "
+			for j in range(0, len(TrainX[i])):
+				Str += str(j + 1) + ":" + str(TrainX[i][j]) + " "
+			Str += "\n"
+		File.write(Str)
+		File.close()
+
+
+
 	print(clf.get_params([True]))
+
+
+
 
 if __name__ == '__main__':
 	if len(sys.argv) != 3:
 		print("Parameter lost")
 		os._exit(0)
 	FileLoc, FileName = Init.GetSufixFile(sys.argv[2], ["png", "jpg"])
+	
+	FileLoc = []
+	FileName = []
+	for i in range(1, 57):
+		FileLoc.append("./Input/PRWInput/Test/Group1/0" + str(i) + ".jpg")
+		FileName.append(str(i))
+	
 	if sys.argv[1] == "test":
 		main("test", FileLoc, FileName)
 	if sys.argv[1] == "train":
